requirements.txt
Django==5.2.6
Pillow==11.3.0
